Reorganize some circuit-related code
- Rename a few circuit-related modules to use the convention `circuit_<category>.rs`.
- Move curve-related targets and methods into a separate module, `circuit_curve.rs`.
