+++ src/lib.rs
@@ -5,14 +5,15 @@
 #![allow(clippy::unreadable_literal)]
 
 pub use bigint::*;
+pub use circuit_bigint::*;
 pub use circuit_builder::*;
-pub use circuit_builder_bigint::*;
-pub use circuit_builder_ordering::*;
+pub use circuit_curve::*;
+pub use circuit_foreign_field::*;
+pub use circuit_ordering::*;
 pub use conversions::*;
 pub use curve::*;
 pub use fft::*;
 pub use field::*;
-pub use foreign_field::*;
 pub use gates::*;
 pub use hash_to_curve::*;
 pub use mds::*;
@@ -30,14 +31,15 @@ pub use verifier::*;
 pub use witness::*;
 
 mod bigint;
+mod circuit_bigint;
 mod circuit_builder;
-mod circuit_builder_bigint;
-mod circuit_builder_ordering;
+mod circuit_curve;
+mod circuit_foreign_field;
+mod circuit_ordering;
 mod conversions;
 mod curve;
 mod fft;
 mod field;
-mod foreign_field;
 mod gates;
 pub mod halo;
 mod hash_to_curve;
